perf: fetch data server-side for create webhooks page
